# openapi/chat-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/files/uploads": {
      "post": {
        "operationId": "Files_createUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/files/{fileId}/complete": {
      "post": {
        "operationId": "Files_complete",
        "parameters": [
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredFileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/files/{fileId}/forward": {
      "post": {
        "operationId": "Files_forward",
        "parameters": [
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForwardFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/files/{fileId}/download": {
      "get": {
        "operationId": "Files_download",
        "parameters": [
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileDownloadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/files/usage": {
      "get": {
        "operationId": "Files_usage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUsageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/files/{fileId}/thumbnail": {
      "post": {
        "operationId": "Files_setThumbnail",
        "parameters": [
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetThumbnailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredFileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/files/{fileId}": {
      "delete": {
        "operationId": "Files_deleteFile",
        "parameters": [
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Chat Server API",
    "description": "Flutter chat backend powered by WuKongIM, LiveKit and S3-compatible storage.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Business access token"
      }
    },
    "schemas": {
      "CreateUploadDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "maxLength": 255
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "type": "number",
            "minimum": 1,
            "maximum": 104857600
          },
          "purpose": {
            "type": "string",
            "enum": [
              "AVATAR",
              "CHAT_IMAGE",
              "CHAT_VOICE",
              "CHAT_VIDEO",
              "CHAT_FILE"
            ]
          },
          "scope": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "DIRECT",
              "GROUP"
            ]
          },
          "scopeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "fileName",
          "mimeType",
          "size",
          "purpose",
          "scope"
        ]
      },
      "ForwardFileDto": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "DIRECT",
              "GROUP"
            ]
          },
          "scopeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "scope",
          "scopeId"
        ]
      },
      "SetThumbnailDto": {
        "type": "object",
        "properties": {
          "thumbnailFileId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "thumbnailFileId"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "statusCode",
          "code",
          "message",
          "requestId",
          "timestamp",
          "path"
        ],
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "code": {
            "type": "string",
            "example": "BAD_REQUEST"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "requestId": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "revokedSessions": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "StoredFileResponse": {
        "type": "object",
        "required": [
          "id",
          "originalName",
          "mimeType",
          "sizeBytes",
          "purpose",
          "scope",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "originalName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "purpose": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "DIRECT",
              "GROUP"
            ]
          },
          "scopeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "thumbnailFileId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "UPLOADED",
              "READY",
              "REJECTED",
              "DELETED"
            ]
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FileUploadResponse": {
        "type": "object",
        "required": [
          "fileId",
          "uploadUrl",
          "method",
          "headers",
          "expiresIn"
        ],
        "properties": {
          "fileId": {
            "type": "string",
            "format": "uuid"
          },
          "uploadUrl": {
            "type": "string",
            "format": "uri"
          },
          "method": {
            "type": "string"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "expiresIn": {
            "type": "integer"
          }
        }
      },
      "FileUsageResponse": {
        "type": "object",
        "required": [
          "usedBytes",
          "quotaBytes",
          "remainingBytes",
          "fileCount"
        ],
        "properties": {
          "usedBytes": {
            "type": "string"
          },
          "quotaBytes": {
            "type": "string"
          },
          "remainingBytes": {
            "type": "string"
          },
          "fileCount": {
            "type": "integer"
          }
        }
      },
      "FileDownloadResponse": {
        "type": "object",
        "required": [
          "downloadUrl",
          "expiresIn",
          "file"
        ],
        "properties": {
          "downloadUrl": {
            "type": "string",
            "format": "uri"
          },
          "expiresIn": {
            "type": "integer"
          },
          "file": {
            "$ref": "#/components/schemas/StoredFileResponse"
          }
        }
      },
...
      "GenericObjectResponse": {
        "type": "object",
        "additionalProperties": true
      }
    }
  }
}